requirements.txt
flask
langchain
python-dotenv
langchain-together

sqlalchemy-iris
langchain-iris
testcontainers-iris
llama-iris
sentence-transformers
fastembed 
openai 
tiktoken
python-dotenv
pandas
ipykernel
gunicorn

pymongo